Rename/move client.APIConstants to netapi.Constants
