+++ src/main/java/com/suse/saltstack/netapi/Constants.java
@@ -1,9 +1,9 @@
-package com.suse.saltstack.netapi.client;
+package com.suse.saltstack.netapi;
 
 /**
  * Constants used in SaltStack API.
  */
-public class APIConstants {
+public class Constants {
     /** Known values for 'eauth' parameter when logging in */
     public static final String LOGIN_EAUTH_AUTO = "auto";
     public static final String LOGIN_EAUTH_PAM = "pam";
